Add files via upload
Compilation of a list of 20 text messages between 2 people. Implement one as a stack and one as a queue and play them back in order (person1 text, person2 text, repeat). Does the conversation make sense? Why or why not.

Change one of the structures to make the text messages appear in the correct order. Explain why this worked.